+++ src/program/side-effects/handle-control-c.ts
